requirements.txt
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
snowflake-connector-python[pandas]>=3.6
python-dotenv>=1.0
